Add examples of the java.util.time API and move other date/time-related examples to a new datesAndText package.
