+++ examples/src/main/java/edu/pdx/cs410J/datesAndText/AroundTheWorld.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.core;
+package edu.pdx.cs410J.datesAndText;
 
 import java.io.*;
 import java.text.*;
