+++ examples/src/main/java/edu/pdx/cs410J/datesAndText/DateDemo.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.core;
+package edu.pdx.cs410J.datesAndText;
 
 import java.text.*;
 import java.util.*;
