+++ examples/src/main/java/edu/pdx/cs410J/datesAndText/DateTimeFormatterExample.java
@@ -1,0 +1,13 @@
+package edu.pdx.cs410J.datesAndText;
+
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class DateTimeFormatterExample {
+
+  public static void main(String[] args) {
+    DateTimeFormatter format = DateTimeFormatter.ofPattern(args[0]);
+    String formatted = ZonedDateTime.now().format(format);
+    System.out.println(formatted);
+  }
+}
