+++ examples/src/main/java/edu/pdx/cs410J/datesAndText/Today.java
@@ -1,4 +1,4 @@
-package edu.pdx.cs410J.core;
+package edu.pdx.cs410J.datesAndText;
 
 import java.util.*;
